Handle mongodb::error::Error::custom properly
MongoDB errors can be a custom variant wrapping another error type; if
this wrapped error is already a dreip_backend::error::Error, we should
unwrap it when converting to that type, rather than further wrapping in
a DbError.

This is particularly relevant when using Session::with_transaction (as
will be introduced in a following commit), as this requires returning
such wrapped errors in some cases.