Update dependencies
This had a few breaking changes that had to be worked around.

Rocket had some trivial changes to cookies and Outcomes. There were also
breaking changes to FromFormField, but the relevant code was dead so it
was simply dropped.

The Amazon SNS SDK now requires setting the behaviour version; we opt
for whatever the latest is as we don't care about the things it controls
(defaults related to timeouts and latency).

Argon2 changed its default config to now consume 2 GiB of memory; we
replace this with our own config based on the recommendations in RFC9106
to ensure sensible memory usage. This results in slightly slower hash
creation and verification overall (but increased security); we avoid any
penalty to the test suite runtime by using a precalculated hash in the
default admin instead of re-hashing the default password each time.

The good news is that both Rocket and the SNS SDK are now stable, so
there should be much less maintenance burden from breaking changes in
the future.